Parse both setup.py and setup.cfg
